+++ packages/plugin-vertical-embed/web/src/toolbar/inline-buttons.tsx
@@ -2,13 +2,14 @@ import { BUTTONS } from 'wix-rich-content-plugin-commons';
 import { getModalStyles, decorateComponentWithProps } from 'wix-rich-content-editor-common';
 import { ReplaceIcon } from '../icons';
 import getModalCustomStyles from './ModalCustomStyles';
-import PostSelectionInputModal from './postSelectionInputModal';
+import VerticalEmbedInputModal from './VerticalEmbedInputModal';
 import { CreateInlineButtons, TranslationFunction, PluginConfig } from 'wix-rich-content-common';
 
 const createInlineButtons: CreateInlineButtons = ({
   t,
   isMobile,
   settings,
+  locale,
 }: {
   t: TranslationFunction;
   settings: PluginConfig;
@@ -16,15 +17,11 @@ const createInlineButtons: CreateInlineButtons = ({
   locale: string;
 }) => {
   return [
-    { keyName: 'alignLeft', type: BUTTONS.SIZE_SMALL_LEFT, mobile: false },
-    { keyName: 'alignCenter', type: BUTTONS.SIZE_CONTENT_CENTER, mobile: false },
-    { keyName: 'alignRight', type: BUTTONS.SIZE_SMALL_RIGHT, mobile: false },
-    { keyName: 'separator1', type: BUTTONS.SEPARATOR, mobile: false },
     {
       keyName: 'replace',
       type: BUTTONS.EXTERNAL_MODAL,
       icon: ReplaceIcon,
-      modalElement: decorateComponentWithProps(PostSelectionInputModal, settings),
+      modalElement: decorateComponentWithProps(VerticalEmbedInputModal, { ...settings, locale }),
       modalStyles: getModalStyles({
         fullScreen: false,
         isMobile,